Fix slow start up times when the asset cache is large.
When the asset cache has many files in it, the main bottleneck during
startup  is `FileAlterationMonitor::start()`, called from
`AppHomeDiskSpaceStatusBar`, `AssetCacheStatusBar` and
`ImageCacheStatusBar`. They use this utility to recursively watch for
filesystem updates so they can update accordingly.

This change avoids the use of a filesystem watcher of any sort, and
instead polls the filesystem at regular intervals (1 minute currently).
This is enabled by a performance improvement to
`AppUtil.getDiskSpaceUsed()` that avoids the use of
`FileUtils.sizeOfDirectory()`.