Implement a basic processor for macro functions
The main annotation added is `@MacroFunction` that marks a Java method as an implementation of a macro function,
optionally specifying the name of the macro function. If the function is additional marked `@Trusted`, then the function
must be called from a trusted context. Other supporting annotations are `@FunctionName` and `@ContextFor`.

The basic concept in implementation is that each Java parameter is mapped to a macro function parameter. In the simplest
case, the macro function value is passed as an `Object` to Java, and the return value from the Java method is
`Object`. In addition, conversion to and from some basic types (`String`, `BigDecimal`, etc.) is supported to improve
the type safety while reducing boilerplate. Some more complex conversion is also possible, e.g., a converter is provided
to support interpreting strings as `Zone` name or GUIDs.

Some parameters do not map to macro function parameters but can be included in the Java method in case they are
required. They are:
- Parameters with type `Parser` receives the macro parser instance as passed to `AbstractFunction.childEvaluate()`
- Parameters with type `VariableResolver` receive the variable resolver as passed to `AbstractFunction.childEvaluate()`
- `String` parameters marked with `@FunctionName` receive the name of the macro function, as passed to
  `AbstractFunction.childEvaluate()`.

There is one more annotation: `@ContextFor`. This one has not been properly implemented yet. The idea is that one
parameter can be used to provide a context for interpreting another. The common example is when looking up
tokens. Ideally, we want the Java method to be able to use `Token` in the method signature, but doing that lookup in
general requires a `Zone` to restrict the search to. The way we do this is by marking a `Zone` parameter as
`@ContextFor` the `Token` parameter. Meanwhile, overloads without the `Zone` can be considered as defaulting to the
current zone.