Make VOFoW rendering more robust
This solves a case where the edge of the visible area is exactly at the edge of the token's bounds. Depending on
precision, the visible area and the bounds might be considered to intersect, despite actually being disjoint. We now
shrink the bounds prior to checking, so that any detected intersections are due to a real intersection.

We also no longer repeat checks for whether or not to clip a token. Instead we do it once, and set a clip on the
`Graphics2D` if needed. Afterwards, we rely on the clip to exclude rendering as needed. In particular, we no longer
check whether a token is completely outside the visible area, but instead just render. This avoids another precision
issue where the intersection checks may not be accurate.