Let's go over the ideas for this branch.
Concept
=======

The idea here is that we actually don't multithread the model at all, but we also don't require thread confinement. How
is this possible?

We'll take a page out of ORM/entity management/DDD design: access to model types is controlled by repositories. Really,
we also have an entity manager that serves as a context for operations, including supplying a unit of work. From the
entity manager, we can grab a repository that supports the necessary operations to retrieve entities of the particular
type (always supports lookup by ID, but may also support other queries).

When an entity is retrieved from a repository, what actually happens is a _copy_ of the entity is returned. Or, to put
it another way, we don't care what the storage for the entity actually is - it could be a Java object, DB storage,
etc. Changes are thus only reflected in that local copy we have. At the same time, looking up an entity a second time
will always return the same entity, even if looked up by different means (i.e., we use an Identity Map).

In order to save our changes, we must commit the unit of work. To accomplish the persistence, we go through every loaded
object, check if it is dirty, and then write back the changes accordingly. The write back happens in accordance with the
needs of the storage, e.g., DB queries, regular Java writes that respect thread confinement, etc.

In order to really make this pop, we need one more thing: transactions. Yes, that means an operation's changes are
communicated to other actors as a whole, and cannot be viewed partially. This also means that once an operation
starts (e.g., a macro begins), the state of the world has to somehow be "frozen".

Benefits
--------

Once we don't have to worry about threading, we can directly expose the object model to more than just MTScript. Indeed,
all HTML5 contexts and GraalJS contexts can directly access the model.

Alternatives
------------

This system obviously relies on transactional updates, including snapshotting. That could be a lot of extra work and
overhead, especially without completely revamping our model.

**Thread-safe model**

Instead we could aim for a completely thread-safe model. This adds a burden on the programmer, and does not prevent
certain logical errors, such as two actors simultaneously updating pieces of the object model. Sure, thread safety will
make sure things don't break, but the reality is that interleaving reads and writes between actors is still not
correct.

**Active Object**

In this alternative, updates do not take effect immediately, but are queued for writes. The object itself maintains a
queue of updates which it services as it is able to.

The downside here is that reads may or may not reflect intermediate writes. That will certainly cause issues in some
cases as we expect a get to reflect prior sets. This can be solved in a similar way as the main concept, where we hand
out "proxy" objects that are updated in-place, but which also queue a write-back to the main model. This concept differs
in not requiring a unit-of-work or transactions, however an Identity Map may still be required to satisfy the
expectation that multiple lookups return the same object to the caller.