Concept
We have some things in our data model which depend on the current state of the client. Especially we depend on which
zone is current in order to do our stuff, and sometimes even require access to the zone _renderer_ that is current. In
addition to just being a poor separation of concerns, it also means we can't do interesting things with our model such
as run it headlessly without any renderers.