Easier CodeTimer access via ScopedValue
It is now possible to grab the current CodeTimer via `CodeTimer.get()` and just use it. Meanwhile, a timer can be
created wiwth `CodeTimer.using(String, Consumer<CodeTimer>)` to use a particular timer for a scope. The created timer
will be used for the duration of the scope, and will have its results reported at the end.

This allows easy setup and teardown of timing results, will also enabling us to instrument parts of the code that don't
have a natural timer to use. In fact, there is a pattern to follow here, which is that certain large-scale algorithms or
actions will set up a timer, but other parts of the code can simple get the timer and instrument themselves without
worrying about how to find the right timer. _Note that this could have also been done by introducing static start/stop
methods to avoid even needing to get the timer, but that is minutiae_.

TODO ScopedValue is still a preview feature. We can do a poor-man's version by using a thread-local stack of values.