Revamp the AreaTree representation
The entire suite of `Area*` types is now semi-immutable: once an AreaTree is constructed, it will not be
changed. `AreaMeta` in particular does not have to maintain mutable state to progressively build up its vertices. In
fact, it's representation no longer contains an AWT `Area`, but is solely composed of a ring of JTS `Coordinate` (plus
other fields derived from the ring). `AreaMeta`'s containment algorithm directly uses this ring, resulting in a very
efficient comparison.

The `AreaTree` digest algorithm has been simplified by relying on JTS to polygonize our input. This saves us from having
to consume a `PathIterator`, deduplicate points, and making the points precise. JTS will even associate oceans to parent
islands for us. All we have to do is associate the islands to their parent oceans and we're good to go.

The algorithms on `AreaTree` have been simplified a great deal. `AreaContainer` no longer sports inherently recursive
methods that alternate between `AreaIsland` and `AreaOcean`. Rather, it only has `contains()` methods that check itself,
but no children (one overload for points, another for other rings). `AreaTree` is now the one responsible for recursing,
which it does via a loop that gives better nullability proofs.

On that note, `AreaTree.getContainer(Point2D)` has been retired in favour of `AreaTree.get(Coordinate)` that returns a
more complex result. The result is a triple, containing at least the nearest ocean (which may be the deepest container,
or else the parent of it), as well as that ocean's parent island (if there is one) and the island that the point is
located in (if there is one). This greatly simplifies `VisionBlockingAccumulator` as each VBL type follows the same
basic pattern that is more easily expressed with this result type.

Meanwhile, AreaContainer now sports methods that were common between AreaOcean and AreaIsland, but used to be spelled
different. Thanks to the inlusion generic parameters for the self and child types, these methods can be typed more
generally.

Better names

Generic cleanup, add a method

Cleanup: only store the Coordinate origin in VisionBlockingAccumulator