Revamp the AreaTree representation
The main structural change here is that `AreaTree` is much more cohesive, allowing it to control more of its own
behaviour. The public interface no longer permits modifying the tree, and `AreaTree` will not modify itself after it is
built. `AreaTree.digest()` is now a public static factory method instead of a pseudo-constructor, which reads better
both internally and externally. `AreaContainer`, `AreaIsland`, `AreaOcean`, and their mutually recursive algorithms are
gone now in favour of a single small `Node` type that does nothing but reference its children and its associated
region (`AreaMeta`), and can only be modifed by `AreaTree. `AreaTree` itself now provides any algorithms that are
required for the tree.

`AreaTree.digest()` is more performant and straightforward than it used to be due to the use of JTS' `Polygonizer`. It
frees us from having to consume a `PathIterator`, enforce precision, and deduplicate points. And most importantly, it
associates oceans with their parent islands for us, leaving us with the much simpler problem of just associated islands
with their parent oceans.

`AreaMeta` has been simplified as well. It is no longer mutable, and doesn't posses or build an AWT `Area`. It is
represented as a plain list of `Coordinate` defining the polygonal boundary of the region it represents. The other
fields exist to speed up specific operations, but in principle could be rmeoved in favour of direct use of the
vertices. `AreaMeta` is also now capable of representing the root ocean, which simplifies `AreaTree` nodes as they no
longer have to worry about possibly-null `AreaMeta` just for that one case.

One more note on algorithms: `AreaTree.getContainer(Point2D)` is now `AreaTree.locate(Coordinate)` which returns a
`LocateResult` triple. This result type will always contain the deepest containing ocean, and will also have that
ocean's parent island and the child island that the point is located in (if such islands exist). This beneifts
`VisionBlockingAccumulator` as it no longer needs to traverse up the tree, and can avoid some of the case work
inherent with that.

Finally, `AreaTree` now keeps track of which type of topology it was built for. This improves the interface of
`VisionBlockingAccumulator` since it doesn't need separate methods for each topology type, but can instead just receive
an `AreaTree` and figure out what to do with it.