Tweak vision algorithm to use with LineSegment instead of LineString
The basic algorithm is still the same, but logic in both `AreaMeta` and `FogUtil` can be simplified by only working with
the individual segments rather than strings of segments. Also introduced a new type `VisionBlockingSet` to abstract the
collection of these segments, which will enable easier representational changes in the future as well.

As a result of this change, `AreaMeta` no longer has to build any geeomtry, and so not loner accepts a
`GeometryFactory`.

Do not bother with preparing vision geometry; all we need from it these days is the bounds, which are easier to compute

Avoid building Geometry in visibility sweep; main reason is that Geometry is complicated and we can avoid that generally