Refine the vision sweep to use  LineSegment instead of LineString
The core of the algorithm is still the same, but logic can be simplified when only dealing with two-vertex `LineSegment`
instead of arbitrarily long `LineString`. It also enables some important checks, especially regarding orientation which
is easily defined for a `LineSegment` unlike a `LineString`. It also means we do not need to use `Geometry` logic, but
can in many cases use more direct and efficient operations on `LineSegment` or its coordinates.

In order to abstract the internal needs of the algorithm from the production of blocking geometry, a new
`VisionBlockingSet` has been introduced that is responsible for taking a segment string (in the form of a list of
vertices) and transforming it into a form acceptable for the sweep algorithm (in this case, line segments).

Another shift away from `Geometry` is the use of the bounding box of the original vision area rather than performing
precise checks against the vision geometry. There is a trade-off here as the more precise we can be the smaller the
input to the algorithm will be, but building and querying the geometry is expensive. In our case the original vision
area tends not to be complicated, so the bounding box is a fairly good approximation with not much being lost.