Revamp vision sweep to avoid wasteful ray intersections
This change moves much closer to Asano's original algorithm as described in section 3.2 of "Efficient Computation of
Visibility Polygons" ([arXiv:1403.3905](https://arxiv.org/abs/1403.3905)). A critical change is in how we find the
nearest open wall, in particular by avoiding expensive intersection calculations on an unordered collection. Instead we
now maintain an ordered collection of open walls, so that querying the closest one is as simple as getting the first
element.

The implementation has been moved out of static `FogUtil` methods and is now encapsulated in `VisionBlockingSet`. As
segments are added, some preparatory work can be done such as expanding the bounding box, deduplicating endpoints,
marking endpoints as starting and ending associated walls, and building the initial open wall set. Once all segments are
added, the `solve()` method will do the finishing touches and perform the sweep itself, producing a polygonal ring.

Some details on the implementation:
1. Open walls are stored in a `TreeSet<>` rather than a flat list. This allows us to find the current nearest wall in
   O(log n) time vs O(n) time, while also supporting efficient removal of ended walls.
2. Intersections with walls are only calculated when needed. These are expensive operations, and we can instead achieve
   sorting using only 3 relatively cheap orientation tests - _in the paper, 5 are mentioned, but two of them we handle
   by requiring a consistent orientation when providing the walls_.
3. We only do one nearest wall lookup per iteration, instead of the two that we did previously. It was unavoidable
   before, since we convolved our nearest wall lookup with our interesection test, and might have needed the
   intersection point for either wall. But now that those two are separate, we can just remember the nearest wall from
   the previous iteration and then perform intersections on whichever wall is needed.
4. Endpoints are sorted using a cheaper and more precise comparison method. The previous one used a custom stand-in for
   polar angle which could then be used for sorting. This was relatively expensive due to the normalization required,
   and the result had to be stored in the endpoint in order to remain efficient. The new method relies on JTS
   orientation and can be computed on demand with no noticable change in performance. Endpoints are now just a position
   and a list of started and ended walls.
5. VisionBlockingSet is poolable, to avoid constant reallocation of the data structures.