Poolable VisionBlockingSet
It's not just a container of segments anymore, it is the vehicle for the vision blocking algorithm. As segments are
added, it builds up the necessary graph structure, hopefully in an efficient manner. The major concerns are the many
hash table lookups on Coordinate, and the need to sort the entire vertex space at the end.

When the set has all the segments, the `.solve()` method will augment it with a bounding box and compute a visibility
polygon. The set can than be reused (pooled) by using the `init(Coordinate)` method to start a new calculation at a new
place.

Doing this allows us to make a few more intelligent decisions. For instance, we don't need to copy our `HashMap<>`
values just to append one element. We also don't need to keep the list sorted until required, which is the most
efficient way of sorting.