Factor out common VBL logic in VisionBlockingAccumulator
By the nature of VBL trees as alternating islands and oceans, and since vision in only blocked at the boundaries of
each, all VBL types follow a common pattern, although it used to be obscured by a lot of case work:
1. Find a container that vision is allowed to pass through (oceans for Wall and Cover, islands for Hill and Pit).
2. Use the container's boundary to block vision.
3. Use the container's childrens' boundaries to block vision.

The new `blockVisionBeyondContainer(AreaIsland)` and `blockVisionBeyondContainer(AreaOcean)` helpers make this pattern
clear, with the different implementations only having to decide which containers should take part. The `AreaIsland`
overload was previously called `addIslandForHillBlocking(AreaIsland)`, but the new naming makes it clearer that it is
not specific to Hill VBL and is consistent with the `AreaOcean` version.

Case work has also been flattened for Hill VBL and Cover VBL by recognizing that all results from
`AreaTree.getContainerAt(Point2D)` either be an ocean or have an ocean as a parent. Additionally, the case work is
identical in both cases. So now we anchor the implementations around the ocean (whether a parent or not), and then mix
in the presence of islands as needed.