Better visilibty graph representation
Before we would add `LineSegment` to `VisibilitySweepEndpoint` in order to represent started and ended walls. But if we
need to do something with the other end, we would have to look up the coordinate in a hash table. This expense is now
avoided by having `VisibilitySweepEndpoint` directly reference each other, by representingg a started or ended wall as
the endpoint on the other end of the wall.

We still need to represent open walls as real line segments, but we now instantiate `LineSegment` only when a wall is
added to the open set.