Better endpoint deduplication
Before we would cautiously add all created `VisibilitySweepEndpoint` to a hash table to make sure we never accidentally
created two of them for the same position. While this works, it ignores the practical reality that duplicates are very
rare, and therefore it is not good that we have to pay a relatively heavy price for every endpoint.

In the new approach, we don't worry about creating duplicate `VisibilitySweepEndpoint`, but instead just add new
instances to the endpoint list for every point we encounter. Once all the segments have been added, we then sort the
list (as we always have), then can easily merge duplicates together since they will all sit next to each other in the
list. The obsolete duplicates are removed by nulling them out, so that we can avoid costly element shifts in the list.