Add occlusion culling
There are plenty of cases where an endpoint encountered in the sweep algorithm is hidden behind other open walls, and
therefore doesn't itself contribute to the result. We must generally still add its started walls, as those can "peek
out" and have an impact, but we can fast-track endpoints like this to avoid some of the core sweep logic. Even when new
open walls have been added for this endpoint, we know none of them will be nearest than the current nearest wall, so we
can skip any attempt to detect changes there.

As a special case of the above, some of the endpoint's started walls might by completely hidden behind the current open
walls. In this case we don't even have to add the wall to the open wall set, saving even more on expensive set
operations.

While this could be expand to check against all current open walls, we currently only check against the nearest open
wall as that still gives good results without too much iteration.