Functional light fade
`DrawableLight` has been liberated from `DrawablePaint` which was designed for a different purpose. And `DrawableLight`
is now responsible for picking the paint, whether a FLAT color or radial FADE. It also accounts for the light center and
radius, something `DrawblePaint` could be extended to do but would just be cruft for anything other than the radial
paint.

Alpha blending for our `BlendingLights` implementation is changed to incorporate both source and destination
alphas. This make allows fades to use an alpha-based fade which looks much better than fading to solid black when using
the Overtop lighting style.