Rework SocketConnection to be in between the new and old approach
Like develop, we do not eagerly create `Socket` when connecting by hostname and port. Instead, `.open()` must be
called to create the socket.

Unlike develop, the `String, Socket` constructor _does not_ implicitly call `.open()` anymore. We require the server to
open any created connections.

This specific combination allows us to not worry about whether the send thread has been set, since it's the caller's
fault if it isn't. At the same time, it means we can create the connection option without it actually trying to connect
to anything. This also means the constructors do not throw `IOException` anymore, which is a simplification that bubbles
up a fair ways.