Change ClientMessageHandler to prefer working with its client
Contrast with prior to this change where everything went through `MapTool.*'` methods. There are still uses of the them,
but they are basically constrained to `MapTool.getFrame()` and player lists (which haven't been merged into the client
yet). There is no use of `MapTool.getCampaign()`, `MapTool.getPlayer()`, or `MapTool.setServerPolicy()` in
`ClientMessagerHandler` anymore.