Clean up PlayerDatabase.getOnlinePlayers()
To start, `LoggedInPlayers.getPlayers()` no longer copies its set, but returns an unmodifiable set. In turn, the
`PlayerDatabase.getOnlinePlayers()` implementations are now consistent in not copying or wrapping the result. Also none
of them throw exceptions, and the declared set of exceptions are pretty random and so have been removed.