Split MapToo.disconnect() between stopServer() and closeClient()
The key internal benefit is that `MapToolClient` state determines whether closure is expected or not, without the need
to signal expected closures to the `ServerDisconnectHandler`. Instead, unexpected disconnects are detected by checking
whether the client was previously closed; if it was, it is expected. This was already true implicitly, but now it is
baked in.

The key external benefit is that we no longer apply server disconnection logic when the only the client needs to be
closed. In particular this means only one "You have disconnected." message is sent, whereas it used to be three for the
host and two for everyone else. And - whether due to this change or a previous one I don't know - we also only get one
error message when trying to start a server on an already-bound port; moreover, application state is correctly restored
to a persoanl server, i.e., it no longer functions like it is connected to the non-existent real server.