Remove impotent static initializer from `Players`
Since `PasswordFilePlayerDatabase` is the only implementation of `PlayerDBPropertyChange`, the instanceof check in
`Players` static initializer can only succeed if a real server based on password files has been started prior to
`Players` ever being loaded... which is impossible. So it has been removed.

With that, it was became obvious that the property change code afforded by `PlayerDBPropertyChange` was also never used,
so the entire interface has been removed. `PasswordFilePlayerDatabase` no longer need proeprty change support. Also
noticed that the way it worked was a carbon copy of the logic in `Players`, so I appreciate the reduced overlap
duplication.

***Note for the future:** Yes, it's possible we may want `PasswordFilePlayerDatabase` to pump events at some point, say
if the file is modified by another MT instance. But today is not that day! And if it ever is, `PropertyChangeSupport` is
not likely to be the correct mechanism, rather the event bus should be used for this.