Revert "Rename MapToolClient to MapToolConnection"
This reverts commit 1d5d7ea3d1281e351d691c609b8880e49b5e977e.

Basically I think the dichotomy between `MapToolServer` and `MapToolClient` is more important than
`MapToolServerConnection` and `MapToolConnection`. At the same time, we definitely do not need both `MapToolClient` and
`MapToolConnection`, so those will stay merged together. In fact, now I'm wondering whether `MapToolServer` and
`MapToolServerConnection` need a similar rethink ... the split between the two seems very arbitrary, and
`MapToolServerConnection` is not a connection to the server, nor a server-side connection, but rather a listener for
managing a mapping from players to server-side client connections.