BAD Tried to implement ServerObserver.serverFaulted(), but this is bad on so many levels
First of all, the server should be capable of closing itself down when this happens, only relying on a callback for
out-of-server stuff that needs to happen. Second (not a problem with this change, but in general), things like the LAN
announcer and UPnP cleanup should be part of the server code proper. Finally, the big issue is that this "corrected"
case is only possible if the server was previously closed with a call to `MapTool.stopServer()`, so calling it again is
an aboslute no-no. In fact, in this case, we should just be able to stop trying to reconnect and let things fall where
they may (the server should already be on its way out).

I am also noticing that the retry mechanism itself is flawed. It's exponential backoff, but:
1. Unbounded, meaning retries can lead to arbitrarily long wait times
2. Really unbounded, meaning that with enough retries we get integer overflow and a broken protocol (connection will
   never be tried).
3. Implemented with a new thread each reconnect. And rest assured, the connection logic itself is running on the new
   thread each time, meaning we have writes without proper synchronizaiotn. Oh, and we also each time create a brand new
   signaling client _and_ a new web socket client with its own thread. Unsure right now if we can do anything about
   that, but the threading situation is a bit dire.
4. Implemented using thread sleeps. Ah, this explains (3). But we can do better using a scheduling executor rather than
   brand new threads. At the very least use green thread in a case like this I think, but really better not too.
5. The retry mechanism is unsynchronized. Before you say anything, this is different from (3). It is possible for the
   retry mechanism to be waiting (as in (4)), but meanwhile another onClose() message or server stop or soemthing comes
   in. Then we have a thread waiting to do connection stuff while we've already instructed the server to dismantle. We
   need to be careful in cases like that.

We can totally stuff the whole backoff mechanism into a generic and small `Retry` interface, with
`ExponentialBackoffRetry`, `FibonacciBackoffRetry`, etc., which properly cap retries / wait times and which can execute
any code.