Handshake.startHandshake() does not need to throw ExecutionException & InterruptedException
These are internal details specifically to support the use of CompletableFuture for one little thing in
ClientHandshake.startHandshake(). If that method simply didn't use CompletableFuture, there would be no exception to
think of. Not that we can do anything about that right now, but at least we can wrap `ExecutionException` and
`InterruptedException` into `IOException`, something the caller is more likely to expect.

TODO There is actually no reason for `startHandshake()` to be in the `Handshake` interface. If it weren't, the server
code would not also have to deal with these exceptions, and possibly we could rework the client code a bit. In fact, the
`Handshake` interface is a bit wierd since we should always know whether we are dealing with a `ClientHandshake` or a
`ServerHandshake`.