Completely separate ClientHandshake and ServerHandshake
There is no need for a common `Handshake` interface as we always know (indeed, *must* know) which type is being
used. There were also some interface requirements that were specific to one implementation but foisted onto the other
one just to make the types work out a bit more nicely.

Now there is no more `Handshake` interface, unnecessary methods and exceptions were removed, and `HandshakeObserver` is
now generic and does not care about a base type. It is homeomorphic to `Consumer<T>`, just with a name that is more
specific to the current purpose, and keeping it allows us to expand the interface in the future.