Change get*VBL() to return separate polygons
The return value is already a list, but now it is possible to contain more than one element. Previously the list of
points would be mashed into a single polygon despite actually being separate polygons. Also we do not start at some
"default" point anymore as that is generally not a part of any polygon but the first.

Also removes the `consumeAreaPoints` helper function as it's actually more code than just consume a `PathIterator`.