Pre-filter non-visible tokens
Then also make the visible token set depend entirely on this filtering and include stamps in addition to tokens, figures
and over-FOW (VBL) tokens (#3715, #3640). I don't personally care that this isn't backwards compatible because the
existing behaviour is garbage, only serving to expose internal rendering decisions rather than anything meaningful about
the zone.

In filterNotVisibleTokens(), I have notes about how what we really want is checks similar to FOW rendering. I.e., we
shouldn't be checking whether the visible area is empty, but rather what area is considered visible for the purposes of
rendering, which should be the same as or similar to the soft-hard FOW boundary.