Make ServerHandshake2 follow a decomposed State pattern
There are technically some minor changes in behaviour, but most of them are clear bugs or mistakes, and none of them
have a large impact. E.g., sending the random GM password for both the player and GM challenge does not affect
functionality as the client doesn't care about its exact values (except to decrypt it back), it's just the we clearly
intended a separate password for the player challenge - and indeed generated one! Another example is that certain error
states are different - it used to be that any protocol error would set the state to a "player blocked" state, but a real
case of blocking a player was set to a generic "error" state. This has now been sorted out.

A weakness of this implementation is the lack of synchronization. For the most part this will be fine - we can assume
the code runs in the message handling thread for the associated server connection. EXCEPT that isn't entirely true since
the server code itself (running on some other thread) also interacts with the handshake object (especially to add an
observer and grab data back out, which may or may not be on a different thread). Additionally, the nature of the
`AwaitingPublicKeyState` is such that UI interaction is required on the swing thread, with callbacks to react to the
acceptance or rejection of the ky. Those callbacks naturally run on the swing thread as well, meaning they do need to be
sequenced relative to everything else. In practice this actually works out for us, because our state transitions impose
a strict sequencing, provided the client doesn't send random messages that could cause a state transition around the
time of the UI interaction.

A strength of the implementation is clearer fault tolerance. All transitions take place in a catch-all so that any
unexpected errors result in an error state. Additionally, `startHandshake()` does not throw exceptions, instead simply
transitioning to a starting state. If there was an exception, it would be caught and handled, immediately transitioning
the handshake to an error state. This does need to be slightly improved, since an error during error handling can result
in repeated erroring as we try to transition away from the current state.

It is my firm intent to introduce a better threading model for server code. The `MapToolServer` itself will have a
dedicated thread with a message pump. Outgoing messages are likely not to be affected by this, since they simply need
stateless encoding. However, incoming messages will be handled by queueing them into the message pump rather than being
handled on whatever connection thread happened to receive it. This will significantly reduce the risk of race
conditions.

In addition to handling incoming messages from connections, any mutations of the server can also be placed on the
message queue. In particular, this will allow us to address the synchronization concerns around `AwaitingPublicKeyState`
since those callbacks can be scheduled on the server message pump. It remains to be seen, but it is possible that other
state changes in the server code ought to be thusly queued as well. However, it is important to note that we DO NOT want
a future-based API here. Essentially, getters should return whatever state is current, while other mutations should
simply be queued to be executed. If we expect that callers would expect certain mutations to be blocking / waitable,
then what we should actually do is action those right away - such properties would have to be properly synchronized, and
must not be the sort where pending tasks might be affected by such a change. We'll see how much this particular issue
comes up though.

**NB** As an implementation detail, so far I have been picturing a concurrent queue to implement the message pump. But
then I would have to be responsible for reading the message queue, handling thread notifications, possibly having to
account for performance issues. A much better approach would be to have a single threaded `Executor` on which all server
tasks must run. It is trivial to queue up a message on such an executor, and we can rely on the Java devs to make it
work correctly.

In a similar vein, connections themselves currently have their own threads but need to potentially accept interactions
from anywhere. We could try to constrain the call sites, but I think it is likely we should support multithreaded use of
connections. Regardless, there is a lifecycle issue that affects handshakes: if we close a connection, that does not
mean that all messages will have been written out. This is because the socket we are closing does not know about the
pending message queue, or even the writer thread. So it is important that we are able to action closing through the
connection's outgoing message queue. I.e., treat close like a message send, to be actioned only after any other pending
messages are written. Once the special close message is encountered, the socket can be closed. Any messages sent after
the close is requested are ignored.