Remove message channels and spooling
Although this removes the "fair" round-robin channel sending, the reality is that the legacy implementation never had
fairness. Instead it had a really complicated way of doing a linear queue, though the intent clearly was to accomplish
some sort of fairness.

Spooling must be done at the server level, not at the connection level. We might reintroduce spooling again in the
future if it operates at that level.

Since we aren't doing spooling anymore, we have no need to support channel IDs in the connection API.