Remove message channels and improve connection threading
The idea behind channels seems to be that we can fairly share a connection between asset transfers and other messages by
using separate queues and a round-robin selection when pulling from the queue. Unfortunately, the implementation is does
not accomplish this, but instead is actually a FIFO queue, meaning no fairness is applied. While this could be fixed, it
is not a good responsibility for `Connection` which already has plenty on its plate. The implementation also has other
weaknesses, such as not being thread-safe, which means we have to be careful to take locks before pushing or pulling
messages. Altogether, it isn't worth keeping around.

Instead, we now use a single `BlockingQueue`. A message can be read via `AbstractConnection.nextMessage()`, which blocks
for a short time waiting for a message. This simplifies the send thread loops for each connection type as they no longer
have to manage their own wait/notify pairs, and don't need to check if messages are available. Instead they can simply
call `nextMessage()` and handle a message if it is returned.

While updating the connection thread as described in the previous paragraph, there were several other threading-related
issues that were solved, especially:
- Shared field are now atomic (particularly shutdown flags)
- For WebRTC, the connection state is checked through a constained set of operations on SendThread.