Add new types for a more decomposed server
Different connection types are now represented by `Channel` implementations. These are simple duplex byte streams with
no understanding of message structure. One is provided for communicating through sockets, and one for WebRTC. To better
support the WebRTC version, the relevant logic from `WebRTCConnection` was factored into the more encapsulated
`SignalingClient` and `PeerConnection`. One decorator is also provided: `ZstdChannel` wraps the raw byte stream into
compressing/inflating streams.

On top of `Channel`, we now have `ChannelConnection` which can wrap any `Channel` into a `Connection`. This class is
where the threading lives, as well as the logic for chunking and dechunking streams. `ChannelConnection`, along with
the `Channel` implementations, is a complete replacement for `AbstractConnection` and subclasses.

We also have new types for the server. `Router` simply keeps track of connections and allows sending messsages to the
connection (`sendMessage()` and `broadcastMessage()`). It does care where connections come from, so is not tied to any
underlying mechanism. The typeical way for connections to be added to a `Router` is by joining it with a
`ChannelReceiver`. A `ChannelReceiver` is much like a `ServerSocket` - it waits for an incoming connection, and produces
`Channel` instances when such a connection comes in. This can be wrapped in a `ChannelConnection` and passed to the
`Router` for it to be part of the server.

The biggest deviation from the existing framework is that handshakes are not considered part of this low-level
library. That is an application-level concern that can be slotted in between the `ChannelReceiver` and the `Router`.